// src/app/cuenta/pages/configured-car/interface/car.interface.ts
export interface Car {
  // definir las propiedades del coche
  cardId: number;
  id: number | null;
  pack: number | null;
  brand: number | null;
  model: number | null;
  engine: number | null;
  extColor: number | null;
  intColor: number | null;
  tire: number | null;
  user: number | null;
  price: number | null;
}